requirements.txt
Flask
beautifulsoup4
Pillow
lxml
